Componentize pagination
