+++ src/components/admin/admin-listings.tsx
@@ -3,6 +3,7 @@ import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
 import type { RecentListing, PendingListing } from "@/types/admin/listings";
 import { getAdminStats, getRecentlyListedItems, getFirstListingsAwaitingApproval, approveFirstListing, rejectFirstListing, deleteListing } from "@/actions/admin-actions";
+import Pagination from "@/components/ui/Pagination";
 
 export default function AdminListings({ userRole }: { userRole: string | null }) {
     const router = useRouter();
@@ -56,72 +57,6 @@ export default function AdminListings({ userRole }: { userRole: string | null })
 
         load();
     }, [userRole, router, listingsPage, pendingListingsPage]);
-
-    function renderPagination(currentPage: number, totalItems: number, itemsPerPage: number, onPageChange: (page: number) => void) {
-        const totalPages = Math.ceil(totalItems / itemsPerPage);
-        if (totalPages <= 1) return null;
-
-        const getPageNumbers = () => {
-            const pages: (number | string)[] = [];
-            const maxVisible = 5;
-
-            if (totalPages <= maxVisible) {
-                for (let i = 1; i <= totalPages; i++) pages.push(i);
-            } else {
-                pages.push(1);
-                if (currentPage > 3) pages.push("...");
-                const start = Math.max(2, currentPage - 1);
-                const end = Math.min(totalPages - 1, currentPage + 1);
-                for (let i = start; i <= end; i++) pages.push(i);
-                if (currentPage < totalPages - 2) pages.push("...");
-                pages.push(totalPages);
-            }
-
-            return pages;
-        };
-
-        return (
-            <div className="flex items-center justify-between border-t border-gray-200 pt-4 mt-4">
-                <div className="text-sm text-gray-600">
-                    Showing {Math.min((currentPage - 1) * itemsPerPage + 1, totalItems)} - {Math.min(currentPage * itemsPerPage, totalItems)} of {totalItems} items
-                </div>
-                <div className="flex items-center gap-2">
-                    <button
-                        onClick={() => onPageChange(currentPage - 1)}
-                        disabled={currentPage === 1 || loading}
-                        className="px-3 py-1 border border-gray-300 rounded-lg hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed transition text-sm"
-                    >
-                        Previous
-                    </button>
-                    {getPageNumbers().map((page, index) =>
-                        typeof page === "number" ? (
-                            <button
-                                key={index}
-                                onClick={() => onPageChange(page)}
-                                disabled={loading}
-                                className={`px-3 py-1 border rounded-lg transition text-sm ${
-                                    currentPage === page ? "bg-green text-white border-green" : "border-gray-300 hover:bg-gray-50"
-                                } disabled:opacity-50 disabled:cursor-not-allowed`}
-                            >
-                                {page}
-                            </button>
-                        ) : (
-                            <span key={index} className="px-2 text-gray-400">
-                                ...
-                            </span>
-                        )
-                    )}
-                    <button
-                        onClick={() => onPageChange(currentPage + 1)}
-                        disabled={currentPage === totalPages || loading}
-                        className="px-3 py-1 border border-gray-300 rounded-lg hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed transition text-sm"
-                    >
-                        Next
-                    </button>
-                </div>
-            </div>
-        );
-    }
 
     function handleViewListing(listingId: string) {
         router.push(`/market/listing/${listingId}`);
@@ -264,7 +199,13 @@ export default function AdminListings({ userRole }: { userRole: string | null })
                             ))
                         )}
                     </div>
-                    {renderPagination(listingsPage, totalListingsCount, LISTINGS_PER_PAGE, setListingsPage)}
+                    <Pagination
+                        currentPage={listingsPage}
+                        totalItems={totalListingsCount}
+                        itemsPerPage={LISTINGS_PER_PAGE}
+                        loading={loading}
+                        onPageChange={setListingsPage}
+                    />
                 </div>
             </div>
 
@@ -314,7 +255,13 @@ export default function AdminListings({ userRole }: { userRole: string | null })
                             </div>
                         ))
                     )}
-                    {renderPagination(pendingListingsPage, totalPendingListingsCount, PENDING_PER_PAGE, setPendingListingsPage)}
+                    <Pagination
+                        currentPage={pendingListingsPage}
+                        totalItems={totalPendingListingsCount}
+                        itemsPerPage={PENDING_PER_PAGE}
+                        loading={loading}
+                        onPageChange={setPendingListingsPage}
+                    />
                 </div>
             </div>
