+++ src/components/admin/admin-users.tsx
@@ -5,6 +5,7 @@ import { $Enums } from "@prisma/client";
 import type { User, SuspendedUser } from "@/types/admin/users";
 import { getAdminStats, getAllUsers, getSuspendedUsers } from "@/actions/admin-actions";
 import { updateAdminUser, suspendUser, banUser, setUserActive } from "@/actions/user-actions";
+import Pagination from "../ui/Pagination";
 
 export default function AdminUsers({ userRole }: { userRole: string | null }) {
     const router = useRouter();
@@ -70,72 +71,6 @@ export default function AdminUsers({ userRole }: { userRole: string | null }) {
 
         load();
     }, [userRole, router, usersPage, suspendedUsersPage]);
-
-    function renderPagination(currentPage: number, totalItems: number, itemsPerPage: number, onPageChange: (page: number) => void) {
-        const totalPages = Math.ceil(totalItems / itemsPerPage);
-        if (totalPages <= 1) return null;
-
-        const getPageNumbers = () => {
-            const pages: (number | string)[] = [];
-            const maxVisible = 5;
-
-            if (totalPages <= maxVisible) {
-                for (let i = 1; i <= totalPages; i++) pages.push(i);
-            } else {
-                pages.push(1);
-                if (currentPage > 3) pages.push("...");
-                const start = Math.max(2, currentPage - 1);
-                const end = Math.min(totalPages - 1, currentPage + 1);
-                for (let i = start; i <= end; i++) pages.push(i);
-                if (currentPage < totalPages - 2) pages.push("...");
-                pages.push(totalPages);
-            }
-
-            return pages;
-        };
-
-        return (
-            <div className="flex items-center justify-between border-t border-gray-200 pt-4 mt-4">
-                <div className="text-sm text-gray-600">
-                    Showing {Math.min((currentPage - 1) * itemsPerPage + 1, totalItems)} - {Math.min(currentPage * itemsPerPage, totalItems)} of {totalItems} items
-                </div>
-                <div className="flex items-center gap-2">
-                    <button
-                        onClick={() => onPageChange(currentPage - 1)}
-                        disabled={currentPage === 1 || loading}
-                        className="px-3 py-1 border border-gray-300 rounded-lg hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed transition text-sm"
-                    >
-                        Previous
-                    </button>
-                    {getPageNumbers().map((page, index) =>
-                        typeof page === "number" ? (
-                            <button
-                                key={index}
-                                onClick={() => onPageChange(page)}
-                                disabled={loading}
-                                className={`px-3 py-1 border rounded-lg transition text-sm ${
-                                    currentPage === page ? "bg-green text-white border-green" : "border-gray-300 hover:bg-gray-50"
-                                } disabled:opacity-50 disabled:cursor-not-allowed`}
-                            >
-                                {page}
-                            </button>
-                        ) : (
-                            <span key={index} className="px-2 text-gray-400">
-                                ...
-                            </span>
-                        )
-                    )}
-                    <button
-                        onClick={() => onPageChange(currentPage + 1)}
-                        disabled={currentPage === totalPages || loading}
-                        className="px-3 py-1 border border-gray-300 rounded-lg hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed transition text-sm"
-                    >
-                        Next
-                    </button>
-                </div>
-            </div>
-        );
-    }
 
     function handleChange(e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) {
         if (selectedUser) {
@@ -305,7 +240,13 @@ export default function AdminUsers({ userRole }: { userRole: string | null }) {
                             </div>
                         ))}
                     </div>
-                    {renderPagination(suspendedUsersPage, totalSuspendedUsersCount, SUSPENDED_PER_PAGE, setSuspendedUsersPage)}
+                    <Pagination
+                        currentPage={suspendedUsersPage}
+                        totalItems={totalSuspendedUsersCount}
+                        itemsPerPage={SUSPENDED_PER_PAGE}
+                        loading={loading}
+                        onPageChange={setSuspendedUsersPage}
+                    />
                 </div>
             )}
 
@@ -394,7 +335,15 @@ export default function AdminUsers({ userRole }: { userRole: string | null }) {
                         )}
                     </div>
                 </div>
-                {!userSearchQuery && renderPagination(usersPage, totalUsersCount, USERS_PER_PAGE, setUsersPage)}
+                {!userSearchQuery && (
+                    <Pagination
+                        currentPage={usersPage}
+                        totalItems={totalUsersCount}
+                        itemsPerPage={USERS_PER_PAGE}
+                        loading={loading}
+                        onPageChange={setUsersPage}
+                    />
+                )}
             </div>
 
             {/* User Info Modal */}
